requirements.txt
torch
matplotlib
opencv-python-headless
numpy
pycocotools

django
djangorestframework
django-cors-headers